+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Tests/TagsTests.cs
@@ -103,18 +103,19 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Tests
                 [SemanticConventions.AttributeHttpMethod] = "GET",
                 [SemanticConventions.AttributeHttpHost] = "localhost",
                 [SemanticConventions.AttributeHttpHostPort] = "8888",
-                [SemanticConventions.AttributeRpcSystem] = "test"
             };
 
             using var activity = CreateTestActivity(tagObjects);
             activityTagsProcessor.CategorizeTags(activity);
 
             Assert.Equal(OperationType.Http, activityTagsProcessor.activityType);
-            Assert.Equal(6, activityTagsProcessor.MappedTags.Length);
+            Assert.Equal(5, activityTagsProcessor.MappedTags.Length);
+            Assert.Empty(activityTagsProcessor.UnMappedTags);
             Assert.Equal("https", AzMonList.GetTagValue(ref activityTagsProcessor.MappedTags, SemanticConventions.AttributeHttpScheme));
             Assert.Equal("localhost", AzMonList.GetTagValue(ref activityTagsProcessor.MappedTags, SemanticConventions.AttributeHttpHost));
             Assert.Equal("8888", AzMonList.GetTagValue(ref activityTagsProcessor.MappedTags, SemanticConventions.AttributeHttpHostPort));
             Assert.Equal("127.0.0.1", AzMonList.GetTagValue(ref activityTagsProcessor.MappedTags, SemanticConventions.AttributeNetHostIp));
+            Assert.Equal("GET", AzMonList.GetTagValue(ref activityTagsProcessor.MappedTags, SemanticConventions.AttributeHttpMethod));
         }
 
         [Fact]
@@ -167,6 +168,36 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Tests
             Assert.Equal("localhost", AzMonList.GetTagValue(ref activityTagsProcessor.MappedTags, SemanticConventions.AttributeHttpHost));
             Assert.Equal("8888", AzMonList.GetTagValue(ref activityTagsProcessor.MappedTags, SemanticConventions.AttributeHttpHostPort));
 
+            Assert.Equal("value", AzMonList.GetTagValue(ref activityTagsProcessor.UnMappedTags, "somekey"));
+        }
+
+        [Fact]
+        public void ActivityTagsProcessor_CategorizeTags_ExtractsRpcNamespace()
+        {
+            var activityTagsProcessor = new ActivityTagsProcessor();
+
+            IEnumerable<KeyValuePair<string, object?>> tagObjects = new Dictionary<string, object?>
+            {
+                [SemanticConventions.AttributeRpcSystem] = "mysystem",
+                [SemanticConventions.AttributeRpcService] = "myservice",
+                [SemanticConventions.AttributeRpcMethod] = "mymethod",
+                [SemanticConventions.AttributeRpcStatus] = "mystatus",
+                ["somekey"] = "value",
+            };
+
+            using var activity = CreateTestActivity(tagObjects, ActivityKind.Client);
+            activityTagsProcessor.CategorizeTags(activity);
+
+            Assert.Equal(OperationType.Rpc, activityTagsProcessor.activityType);
+
+            Assert.Equal(3, activityTagsProcessor.MappedTags.Length);
+            Assert.Equal(2, activityTagsProcessor.UnMappedTags.Length);
+
+            Assert.Equal("mysystem", AzMonList.GetTagValue(ref activityTagsProcessor.MappedTags, SemanticConventions.AttributeRpcSystem));
+            Assert.Equal("myservice", AzMonList.GetTagValue(ref activityTagsProcessor.MappedTags, SemanticConventions.AttributeRpcService));
+            Assert.Equal("mystatus", AzMonList.GetTagValue(ref activityTagsProcessor.MappedTags, SemanticConventions.AttributeRpcStatus));
+
+            Assert.Equal("mymethod", AzMonList.GetTagValue(ref activityTagsProcessor.UnMappedTags, SemanticConventions.AttributeRpcMethod));
             Assert.Equal("value", AzMonList.GetTagValue(ref activityTagsProcessor.UnMappedTags, "somekey"));
         }